+++ readme/RELEASE_NOTES.md
@@ -1,7 +1,8 @@
 * [#21](https://github.com/skuzzle/snapshot-tests/issues/21): Throw `IllegalArgumentException` instead of `NullPointerException` when validating parameters
 * `TestFile` injected in parameterized tests now has a `sibling(String)` and `siblingWithExtension(String)` method.
+* Added `DirectoriesFrom` which iterates directories and injects them as `TestDirectory`.
 * Added `filter` attribute to the `@FilesFrom` annotation to specify a `PathFilter` implementation. Allows for more fine grained filtering control.
-* Experimental support for changing the snapshot directory per assertion with `snapshot.in(pathToDirectory).assertThat(...)`
+* Experimental support for changing the snapshot directory per assertion with `snapshot.in(pathToDirectory).assertThat(...)`.
 
 Maven Central coordinates for this release: